requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=0.23.0
pytest>=7.0.0
httpx>=0.24.0  # For testing FastAPI
requests>=2.31.0  # For API calls to GitHub/Bitbucket